Simultaneous point picking in multiple point clouds
