+++ src/types.ts
@@ -56,7 +56,6 @@ export interface IPotree {
 export interface PickPoint {
   position?: Vector3;
   normal?: Vector3;
-  datasetNormal?: Vector3;
   [property: string]: any;
 }
